backend/requirements.txt
# Web 框架
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# 大模型 SDK
openai==1.14.0

# 数据处理
pydantic==2.6.4
pandas>=2.0.0

# 数据库
neo4j>=5.0.0

# 向量数据库（RAG）
chromadb>=0.4.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# 文件处理
PyPDF2>=3.0.0
python-docx>=0.8.11

# PDF 导出
reportlab>=4.0.0

# 工具库
python-dotenv==1.0.1
httpx>=0.26.0
langchain-openai>=0.1.0
langgraph>=0.0.40